Fix a few minor Y2K38 printf warnings with time_t
Use intmax_t or uintmax_t instead of a 32 bit integer.
Coverity 1509336, 1509337, 1509342

+++ epan/print.c
@@ -1496,8 +1496,8 @@ print_pdml_geninfo(epan_dissect_t *edt, FILE *fh)
 
     /* Print geninfo.timestamp */
     fprintf(fh,
-            "    <field name=\"timestamp\" pos=\"0\" show=\"%s\" showname=\"Captured Time\" value=\"%d.%09d\" size=\"%d\"/>\n",
-            tmp, (int)edt->pi.abs_ts.secs, edt->pi.abs_ts.nsecs, frame_finfo->length);
+            "    <field name=\"timestamp\" pos=\"0\" show=\"%s\" showname=\"Captured Time\" value=\"%jd.%09d\" size=\"%d\"/>\n",
+            tmp, (intmax_t)edt->pi.abs_ts.secs, edt->pi.abs_ts.nsecs, frame_finfo->length);
 
     wmem_free(NULL, tmp);
